cast to int

+++ src/eigen_rigidbody.hpp
@@ -103,7 +103,7 @@ public:
   }
 
   RigidBodyState(const rigid_body_pose_t * pose) :
-      vec(basic_num_states), utime(pose->utime)
+      vec((int)basic_num_states), utime(pose->utime)
   {
     Eigen::Map<const Eigen::Vector3d> velocity_map(pose->vel);
     Eigen::Map<const Eigen::Vector3d> angular_velocity_map(pose->rotation_rate);
